+++ src/api/pbx.ts
@@ -329,6 +329,7 @@ export class PBX extends EventEmitter {
 
     const as = getAuthStore()
     as.pbxConfig = config
+    as.setRecentCallsMax(config['webphone.recents.max'])
 
     // the custom page only load at the first time the tab is shown after you log in
     //    even after re-connected it, don't refresh it again